refactor: consolidate classic itinerary components and fix scroll trigger
