+++ src/components/pages/home/ClassicItinerary/RecommendationSection/config.ts
