+++ src/components/pages/home/ClassicItinerary/index.tsx
@@ -1,0 +1,5 @@
+export { default as BookShelf } from './BookShelf'
+export { default as BookShelfSection } from './BookShelfSection'
+export { default as DestinationCard } from './DestinationCard'
+export { default as RecommendationSection } from './RecommendationSection'
+export { default as Introduction } from './Introduction'
