+++ src/components/pages/home/HomePageClient.tsx
@@ -1,8 +1,7 @@
 'use client'
 
 import Banner from '@/components/pages/home/Banner'
-import BookShelfSection from '@/components/pages/home/BookShelfSection'
-import RecommendationSection from '@/components/pages/home/RecommendationSection'
+import { BookShelfSection, RecommendationSection } from '@/components/pages/home/ClassicItinerary'
 import CollectionRecommendation from '@/components/pages/home/CollectionRecommendation'
 import Feedback from '@/components/pages/home/Feedback'
 
@@ -10,8 +9,10 @@ const HomePageClient = () => {
   return (
     <div className='min-h-screen bg-figma-neutral-50'>
       <Banner />
+      <div className='relative bg-figma-neutral-50'>
       <BookShelfSection />
       <RecommendationSection />
+      </div>
       <CollectionRecommendation />
       <Feedback />
     </div>
